Implements, add(), addAll(), remove(), removeAll(), and equals()
UserID<?> concrete class fields are replaced with Identifiable interface.
Implements SortedSet<TemporalOccurrence> to allow for multiple occurrences of contact between the same two users.
add(), addAll(), remove(), and removeAll() will be removed in a future commit, but are kept in the case they want to be re-implemented.
equals() claims two Contact instances are equal if both users are the same, regardless of the order they are assigned.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>