Organizes methods in VentilatorReadRequestHandler; simplifies class type; fixes serialization bug
The event type of the WorkerReadRequestHandler only needs to be ContractedPdaRequest. This will allow a lot of the functionality implemented by VentilatorReadRequestHandler to be refactored into a common ventilator handler that can be used by both read and write handlers.

ContractedPdaClient was attempting to write an abstract type as a String; only concrete types implemented by Immutables can be serialized effortlessly with the ObjectMapper.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>