Refactors VentilatorReadRequestHandler; adds VentilatorWriteRequestHandler and WorkerWriteRequestHandler
The core functionality of the ventilator handler is shared between reading and writing; it is only specific worker Lambda functions that differ in what they invoke.

The WorkerWriteRequestHandler retrieves the computed risk score and attempts to write it back to the PDA. If the HAT cannot be found in the output of the algorithm, it is assumed the risk score is 0.0.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>