Allows Contact to have no occurrences
The checking of state to ensure a Contact has at least one occurrence is figured to be more prone to bugs than allowing for no occurrences. As a result, the removal of factor vertices during the first superstep of belief propagation was added back into FactorVertexComputation. With empty occurrences allowed, a small adjustment was made to ContactMatchingComputation such that a null Contact is not returned, but rather just one with empty occurrences.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>