Adds backend.py with updates to algorithm.py and simulate.py
Creates a common interface for an undirected graph with vertex and edge attributes which allows the BeliefPropagation to abstractly interact with this backend. With this interface,
only the backend (i.e. networkx or igraph) needs to be selected. The changes to reflect this interface in algorithm.py and simulate.py have been made.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>