Small change, big performance
Not sure why I did not try this solution before, but using the itertools combination iterator and for_each() with max_concurrency set to the number of physical CPUs performs
insanely well. For example, given 5000 users, there are roughly 12.5 million unique pairs to process. The algorithm does this in (dramatic pause) 0.25 seconds. Not too shabby.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>