Bulk synchronous parallel implementation of BeliefPropagation
The intended advantage of this implementation is to avoid the excessive blocking of using a central VertexStore in remote functions and actors. The factor graph is partitioned
amongst several actors, which maintaining the state of variable and factor vertices. The vertices are assigned in a round-robin fashion to help balance the workload. Each partition
has a local queue that receives messages from the other partitions. A central queue actor is used to receive all outgoing messages from the partitions, which are then routed to the
correct partition. Thus, blocking (i.e. `ray.get()`) is only caused by getting the max scores and variables in the main _call() method of BeliefPropagation.

Unfortunately, this implementation is not seeing an improvement in performance. This may be caused by having a single router queue. However, this implementation will replace the
current implementation since there is less blocking in general and naturally scales with the number of vertices by having additional partitions/actors/CPUs.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>