Checkpoint before refactoring
- Adds AsyncQueue (simple wrapper around asyncio.Queue) to conform to Queue interface
- Makes VertexStore a collections.Collection
- Removes return type hint for VariablePart and FactorPart
- Close-to-testable remote version for BeliefPropagation.

I have decided, in the hope of (1) simplifying the implementation logic and (2) speeding up the remote version by using decorators, I am going to implement separate the local and remote versions independently. This will allow the local version to
leverage the simplicity of its nature while allowing the remote version to not be muddied by unnecessary conditional branching based on the local_mode variable.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>