Checkpoint before reverting to non-queue local implementation
I was not expecting the alternative local implementation to have an impact, but I guess I was wrong. This version runs significantly slower for all numbers of users. Will be reverting back to the original version that iterates over vertices. I
will see if the queue can still stay, but group together the messages being sent by the variables into (what was previously being used) an itertools.chain instance.

After this is fixed, I will move on to the RemoteBeliefPropagation implementation, which will primarily just be a copying over what was already implemented yesterday.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>