Refactors _ShareTraceVariablePart to use asyncio.gather
All aspects that require looping with an await statement inside are refactored so that
all await statements are grouped together inside an asyncio.gather() call. This should
help significantly with regard to performance. Hanging for loop when creating the graph
remains an issue, however.

Allows actors to be detached now. Don't believe it will be an issue since they should
never go out of scope. They will skill be called to kill themselves at the end of run
regardless.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>